add context manager to reset next_variable_name
Create context manager TypeVariableContext which resets the next name
of the type variable. It ensures that every time when we analyse a new
expression, by calling analyse() function inside 'with context()' block, we
start with the name 'a' of type variable. It allows to not rely on the
execution order of integration tests when we define expected type which
contains type variable. In addition, when we call analyse outside 'with'
block and try to access next_variable_name, we will get an error.
See implementation details in lambda_types.

Modify type_inference_test by deleting code which was needed to ensure
that tests are running in the specific order. Change the expected type since
the first type variable name will be 'a' for each test case whose expected
type contains type variable.

Usage example:
with context(): // context() returns instance of context manager
    t = analyse(...)
    print(t) // here next_variable_name will be called